+++ R/MatchedSpectra.R
@@ -91,7 +91,7 @@
 #'
 #' - `addProcessing`: add a processing step to both the *query* and *target*
 #'   `Spectra` in `object`. Additional parameters for `FUN` can be passed *via*
-#'   `...`. See `addProcessing` documentation in [Spectra()] for more
+#'   `...`. See `addProcessing` documentation in [Spectra::Spectra()] for more
 #'   information.
 #'
 #' - `plotSpectraMirror`: creates a mirror plot between the query and each
@@ -99,21 +99,26 @@
 #'   single query spectrum. Setting parameter `scalePeaks = TRUE` will scale
 #'   the peak intensities per spectrum to a total sum of one for a better
 #'   graphical visualization. Additional plotting parameters can be passed
-#'   through `...`.
+#'   through `...`. The parameters `ppm` and `tolerance` can be used to
+#'   define the m/z tolerance for matching peaks between the query and target
+#'   spectra. If not provided by the user, the values from the `param`
+#'   object used to create the `MatchedSpectra` object are used; if these are
+#'   missing, the default values (`ppm =20` and `tolerance = 0`) are used.
 #'
 #' - `setBackend`: allows to change the *backend* of both the query and target
-#'   [Spectra()] object. The function will return a `MatchedSpectra` object with
-#'   the query and target `Spectra` changed to the specified `backend`, which
-#'   can be any backend extending [MsBackend].
+#'   [Spectra::Spectra()] object. The function will return a `MatchedSpectra`
+#'   object with the query and target `Spectra` changed to the specified
+#'   `backend`, which can be any backend extending [Spectra::MsBackend].
 #'
-#' @param backend for `setBackend`: instance of an object extending [MsBackend].
-#'   See help for `setBackend` in [Spectra()] for more details.
+#' @param backend for `setBackend`: instance of an object extending
+#'   [Spectra::MsBackend]. See help for [Spectra::setBackend()]
+#'   for more details.
 #'
 #' @param columns for `spectraData`: `character` vector with spectra variable
 #'   names that should be extracted.
 #'
 #' @param FUN for `addProcessing`: function to be applied to the peak matrix
-#'   of each spectrum in `object`. See [Spectra()] for more details.
+#'   of each spectrum in `object`. See [Spectra::Spectra()] for more details.
 #'
 #' @param main for `plotSpectraMirror`: an optional title for each plot.
 #'
@@ -131,7 +136,7 @@
 #'
 #' @param spectraVariables for `addProcessing`: `character` with additional
 #'   spectra variables that should be passed along to the function defined
-#'   with `FUN`. See [Spectra()] for details.
+#'   with `FUN`. See [Spectra::Spectra()] for details.
 #'
 #' @param target `Spectra` with the spectra against which `query` has been
 #'   matched.
@@ -386,7 +391,7 @@ setMethod("spectraData", "MatchedSpectra",
 setMethod("matchedData", "MatchedSpectra",
           function(object, columns = spectraVariables(object), ...) {
               spectraData(object, columns)
-})
+          })
 
 #' @importMethodsFrom Spectra addProcessing
 #'
@@ -415,6 +420,8 @@ setMethod(
 #'
 #' @importFrom Spectra scalePeaks
 #'
+#' @importMethodsFrom ProtGenerics as.list
+#'
 #' @export
 setMethod(
     "plotSpectraMirror", "MatchedSpectra",
@@ -422,6 +429,12 @@ setMethod(
              scalePeaks = FALSE, ...) {
         if (length(query(x)) != 1)
             stop("Length of 'query(x)' has to be 1.")
+        dots <- list(...)
+        pl <- as.list(x@metadata[["param"]])
+        ppm_res <- .res_setting(dots = dots, param_list = pl,
+                                 name = "ppm", default = 20)
+        tol_res <- .res_setting(dots = dots, param_list = pl,
+                                 name = "tolerance",default = 10)
         y <- x@target[x@matches$target_idx]
         if (!length(y))
             y <- Spectra(DataFrame(msLevel = 2L))
@@ -432,8 +445,19 @@ setMethod(
         par(mfrow = n2mfrow(length(y)))
         for (i in seq_along(y))
             plotSpectraMirror(x = x, y = y[i],
-                              xlab = xlab, ylab = ylab, main = main, ...)
+                              xlab = xlab, ylab = ylab, main = main,
+                              ppm = ppm_res, tolerance = tol_res,...)
     })
+
+#' Helper for ppm and tolerance setting.
+#' @noRd
+.res_setting <- function(dots = list(), param_list, name, default) {
+    if (!is.null(dots[[name]]))
+        return(dots[[name]])
+    if (!is.null(param_list[[name]]))
+        return(param_list[[name]])
+    return(default)
+}
 
 #' @importMethodsFrom ProtGenerics setBackend
 #'